+++ .github/workflows/e2e-playwright-allure_report.yml
@@ -40,12 +40,6 @@ jobs:
         run: |
           echo "LOCALE=${{ github.event.inputs.Locale }}" >> $GITHUB_ENV
 
-      - name: Download Locales
-        run: |
-          curl -L "https://download-directory.github.io?url=https://github.com/DependencyTrack/frontend/tree/master/src/i18n/locales" -o locales.zip
-          unzip locales.zip
-          cp -r locales ./e2e/playwright-tests/setup/
-
       - name: Generate Random UUID for password use
         run: echo "RANDOM_PASSWORD=$(uuidgen -r)" >> $GITHUB_ENV
